jQuery noConflict() issue
For some reason everything loaded after debug kit stopped working. After some digging and
chatting to people on `#jquery` I was told that the method being used is not "normal".

The changes look big but its just indentation, roughly:

	(function($) {
		// var DEBUGKIT = ...
		// DEBUGKIT.$ = jQuery.noConflict(true); changed to DEBUGKIT.$ = jQuery;
		// DEBUGKIT.loader = ...
	})(jQuery);

The jQuery object is passed into the anonomus function and is then available for use. Before
making this change `$` was not available for other `jQuery` plugins loaded after DebugKit