Updated the SQL Log toolbar to use SQL dump fallback
The SQL log assumes that all content connections have been logged to the database. 
Although caching is important, it's also important to include a fallback method if caching isn't setup and no connection is found.
In this case a simple include to the sql_dump element, which according to the docs should be removed from the Layout.