Allow scrolling on small screen
In responsive design the variables tab gets lost from the toolbar when the screen is small. 
This solution is not elegant but it saves wondering "where did my tabs go" on smaller screen sizes.
Some sort of media-query driven popup would be better in the long run. Maybe that is in the works?