Prevent Debugkit from running in production
Running an application with debug mode on AND installing dev dependencies in a
production environment is obviously the developers' fault for not following the
instructions provided. But we can do better than that.

Instead of blaming them when their secrets are exposed to the internet, we can
run some simple checks that can tells us whether or not the aplications is running
visibly to the www.

In the valid cases developers need DebugKit running in producition-like systems (like a
QA environment), they still have the option of `forceEnable`.