Allow 172.16.0.0/12 and IPv6 private ranges
The isSuspiciouslyProduction method() marked every IP address starting
with 192, 127 or 10 as 'safe', including public addresses starting with
192, like 192.1.2.3. The private class B range, starting with 172.16
through 172.31, was missing completely.

Hosts in the 172-range were therefore marked as public, and developers
had to work around this by using the forceEnable flag. This change
fixes that; only IPv4 and IPv6 addresses that are not in private or
reserved ranges, according to filter_var, are now assumed to be public.